refactor: break down biometrics package in sub packages
