+++ app/src/main/java/org/dhis2/data/biometrics/BiometricsConfigApi.kt
@@ -1,6 +1,6 @@
 package org.dhis2.data.biometrics
 
-import org.dhis2.usescases.biometrics.BiometricsConfig
+import org.dhis2.usescases.biometrics.entities.BiometricsConfig
 import retrofit2.Call
 import retrofit2.http.GET
