+++ app/src/main/java/org/dhis2/data/service/SyncDataWorkerModule.kt
@@ -10,8 +10,8 @@ import org.dhis2.data.biometrics.BiometricsConfigRepositoryImpl
 import org.dhis2.data.biometrics.BiometricsParentChildConfigApi
 import org.dhis2.data.biometrics.BiometricsParentChildConfigRepositoryImpl
 import org.dhis2.data.service.workManager.WorkManagerController
-import org.dhis2.usescases.biometrics.BiometricsConfigRepository
-import org.dhis2.usescases.biometrics.BiometricsParentChildConfigRepository
+import org.dhis2.usescases.biometrics.repositories.BiometricsConfigRepository
+import org.dhis2.usescases.biometrics.repositories.BiometricsParentChildConfigRepository
 import org.dhis2.utils.analytics.AnalyticsHelper
 import org.hisp.dhis.android.core.D2
