+++ app/src/main/java/org/dhis2/usescases/biometrics/entities/BiometricsConfig.kt
@@ -1,4 +1,4 @@
-package org.dhis2.usescases.biometrics
+package org.dhis2.usescases.biometrics.entities
 
 data class BiometricsConfig(
     val orgUnitGroup: String?,
