+++ app/src/main/java/org/dhis2/usescases/biometrics/repositories/BiometricsConfigRepository.kt
@@ -1,4 +1,6 @@
-package org.dhis2.usescases.biometrics
+package org.dhis2.usescases.biometrics.repositories
+
+import org.dhis2.usescases.biometrics.entities.BiometricsConfig
 
 interface BiometricsConfigRepository {
     fun sync()
