+++ app/src/main/java/org/dhis2/usescases/biometrics/repositories/BiometricsParentChildConfigRepository.kt
@@ -1,4 +1,4 @@
-package org.dhis2.usescases.biometrics
+package org.dhis2.usescases.biometrics.repositories
 
 interface BiometricsParentChildConfigRepository {
     fun sync()
